+++ examples/talk-llama/llama.cpp
@@ -1416,7 +1416,7 @@ struct llama_tokenizer {
 
             if (token == vocab_.token_to_id.end()) {
                 // output any symbols that did not form tokens as bytes.
-                for (int j = 0; j < (int) symbol.n; ++j) {
+                for (size_t j = 0; j < symbol.n; ++j) {
                     llama_vocab::id token_id = static_cast<uint8_t>(symbol.text[j]) + 3;
                     output.push_back(token_id);
                 }
